requirement.txt
pandas
requests>=2.25.1
apache-airflow-providers-common-sql>=1.3.1
apache-airflow-providers-microsoft-mssql>=3.6.1
pymssql>=2.1.8
sqlalchemy
pyodbc
boto3
s3fs
apache-airflow[amazon]